Move classes to `Pipeline` namespace
