+++ src/Pipeline/Cleanup.php
@@ -3,7 +3,7 @@
  * Deletes source files and empty directories.
  */
 
-namespace BrianHenryIE\Strauss;
+namespace BrianHenryIE\Strauss\Pipeline;
 
 use BrianHenryIE\Strauss\Composer\ComposerPackage;
 use BrianHenryIE\Strauss\Composer\Extra\StraussConfig;
