+++ src/Pipeline/DependenciesEnumerator.php
@@ -3,7 +3,7 @@
  * Build a list of ComposerPackage objects for all dependencies.
  */
 
-namespace BrianHenryIE\Strauss;
+namespace BrianHenryIE\Strauss\Pipeline;
 
 use BrianHenryIE\Strauss\Composer\ComposerPackage;
 use BrianHenryIE\Strauss\Composer\Extra\StraussConfig;
