+++ src/Pipeline/FileEnumerator.php
@@ -5,7 +5,7 @@
  * Also record the `files` autoloaders.
  */
 
-namespace BrianHenryIE\Strauss;
+namespace BrianHenryIE\Strauss\Pipeline;
 
 use BrianHenryIE\Strauss\Composer\ComposerPackage;
 use BrianHenryIE\Strauss\Composer\Extra\StraussConfig;
